Configuracion y cambios generales

+++ SolucionTarea2/Inc/GPIOxDriver.h
@@ -37,6 +37,6 @@ typedef struct{
 void GPIO_Config (GPIO_Handler_t *pGPIOHandler);
 void GPIO_WritePin (GPIO_Handler_t *pPinHandler, uint8_t newState);
 void GPIOxTooglePin(GPIO_Handler_t*pPinHandler);
-uint32_t GPIO_ReadPin(GPIO_Handler_t *pPinHandler);
+uint8_t GPIO_ReadPin(GPIO_Handler_t *pPinHandler);
 
 #endif /* GPIOXDRIVER_H_ */
